requirements.txt
torch==2.0.1
RealESRGAN==1.0
numpy==1.25.2
Pillow==9.2.0
paramiko==3.3.1
pandas==2.0.3
requests==2.28.1
